fix: gravitational timestep
The gravitational part of the time step should always be positive but
becomes negative when the gravitational potential is negative.  This can
happen if a sufficiently large fixed background potential is added.

This part of the time step should be such that the phase at any point
never changes by pi during a given time step.  A negative potential
would change the phase in the opposite direction, so should be taken
into account.

Add an absolute value to the time step.

Fixes #31