fix: Only calculate the summary in inner loop
Calculating the summary both inside and outside the loop in
`take_steps!` is inconsistent.
This shows up as discontinuities in quantities such as the energy
because the kinetic and gravitational energy are out of phase by half a
time step.

The fix: calculate summaries only in the inner loop.  Any summary
quantities that depend on the phase should be calculated carefully,
potentially with a temporary update to psi by half a time step.

BREAKING CHANGE: summary is calculated at slightly different times