feat: Allow soliton init on other grid types
Soliton initialization is different if the soliton is going to be on an
Array or a PencilArray.  The latter needs one to transform to global indices. As written, the check was against the type of psi_x's container, rather than psi_x itself.

Generalize this.  This allows for initialization of any grid type that
stores psi_x in a Array or PencilArray.