fix: outer_step! call at end of loop missed
The symmetrized split step method combines most calls to `outer_step!` with
half a time step into full time steps.  However, an extra half step has
to be taken each time the time step is updated. This is missing for the
external states.

With default settings, this results in an approximately 5% error.

Add it in.