fix: Correct scale factor on last half step
There is a missing `; a = a(t)` in the argument to `outer_step!` that
does the closing half step.  This is never an issue if there is no
background expansion.

Add it in.

This should maybe not be an optional argument; it shouldn't be possible
to accidentally call with the wrong `a(t)`.