fix: Typo in gravitational energy
9b17f3037c4bb013c1aa714cf51383391314d193
(https://github.com/musoke/UltraDark.jl/pull/152) introduced a bug in
the gravitational energy.
The arguments of E_gravity_density are reversed.  This doesn't show up in
tests because the psi used is real, so the incorrect energy density can
be coerced to be real.

Fix this and update the tests of summaries so that psi is complex.