doc: Link to external objects
Use `DocumenterInterLinks.jl` to link to objects in `Julia.Core`,
`PencilArrays` and `PencilFFTs`.

The docs for the stable branches of `PencilArrays` and `PencilFFTs`
don't have `DocumenterInterLinks` inventories yet, so link to the docs
for the dev branches.

Julia does not have an inventory file at all, so use the one from
https://github.com/JuliaDocs/DocumenterInterLinks.jl/wiki/Inventory-File-Repository

Fixes: https://github.com/musoke/UltraDark.jl/issues/184